move coverage merging into projects
